soldes maintenant en flottant, calcul de cout de troncon avec requete

+++ VerbiageVoiture/src/main/java/fr/verbiagevoiture/vue/Solde.java
@@ -137,7 +137,7 @@ public class Solde {
 	}
 	
 	protected boolean Recharger() {
-		return myco.RechargerSolde(Integer.parseInt(monSolde.getText()));
+		return myco.RechargerSolde(Float.parseFloat(monSolde.getText()));
 	}
 	
 	protected String AfficherSolde() {
